api/requirements.txt
numpy==1.23.5
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
torch==2.0.1 --index-url https://download.pytorch.org/whl/cpu
openai-whisper==20230314